Convert main to TS
